Ensure that RESTEasy Reactive violation report doesn't break Hibernate Validator in native
Fixes: #21516
